requirements.txt
Flask==2.3.3
numpy==1.24.3
opencv-python==4.8.0.76
tensorflow==2.13.0
Pillow==10.0.0
